Debugged
Debugged a bug where the program continued to run wherever it shouldn't when there was an error regarding opening a file. Also, a try-except block is added to import the tabulate, so if a user doesn't have it, a simpler method will be used to display data instead of a tabular form.